+++ hardware/MQTT.h
@@ -38,7 +38,7 @@ private:
 	bool ConnectIntEx();
 	void WriteInt(const std::string &sendStr);
 	void ProcessMySensorsMessage(const std::string &MySensorsMessage);
-	void SendDeviceInfo(const int m_HwdID, const unsigned long long DeviceRowIdx, const std::string &DeviceName, const unsigned char *pRXCommand);
+	void SendDeviceInfo(const int m_HwdID, const uint64_t DeviceRowIdx, const std::string &DeviceName, const unsigned char *pRXCommand);
 protected:
 	std::string m_szIPAddress;
 	unsigned short m_usIPPort;
